Refactor fixtures.
