+++ tests/Process/test_single_point_energy.py
@@ -5,8 +5,8 @@ import BioSimSpace as BSS
 from tests.conftest import url, has_amber, has_gromacs
 
 
-@pytest.fixture(scope="session")
-def system():
+@pytest.fixture(scope="module")
+def ubiquitin_system():
     """Re-use the same molecuar system for each test."""
     return BSS.IO.readMolecules(
         [f"{url}/ubiquitin.prm7.bz2", f"{url}/ubiquitin.rst7.bz2"]
@@ -17,17 +17,19 @@ def system():
     has_amber is False or has_gromacs is False,
     reason="Requires that both AMBER and GROMACS are installed.",
 )
-def test_amber_gromacs(system):
+def test_amber_gromacs(ubiquitin_system):
     """Single point energy comparison between AMBER and GROMACS."""
 
     # Create a single-step minimisation protocol.
     protocol = BSS.Protocol.Minimisation(steps=1)
 
     # Create a process to run with AMBER.
-    process_amb = BSS.Process.Amber(system, protocol)
+    process_amb = BSS.Process.Amber(ubiquitin_system, protocol)
 
     # Create a process to run with GROMACS.
-    process_gmx = BSS.Process.Gromacs(system, protocol, extra_options={"nsteps": 0})
+    process_gmx = BSS.Process.Gromacs(
+        ubiquitin_system, protocol, extra_options={"nsteps": 0}
+    )
 
     # Run the AMBER process and wait for it to finish.
     process_amb.start()
@@ -57,23 +59,25 @@ def test_amber_gromacs(system):
     has_amber is False or has_gromacs is False,
     reason="Requires that both AMBER and GROMACS are installed.",
 )
-def test_amber_gromacs_triclinic(system):
+def test_amber_gromacs_triclinic(ubiquitin_system):
     """Single point energy comparison between AMBER and GROMACS in a triclinic box."""
 
     # Swap the space for a triclinic cell (truncated octahedron).
     from sire.legacy.Vol import TriclinicBox
 
     triclinic_box = TriclinicBox.truncatedOctahedron(50)
-    system._sire_object.setProperty("space", triclinic_box)
+    ubiquitin_system._sire_object.setProperty("space", triclinic_box)
 
     # Create a single-step minimisation protocol.
     protocol = BSS.Protocol.Minimisation(steps=1)
 
     # Create a process to run with AMBER.
-    process_amb = BSS.Process.Amber(system, protocol)
+    process_amb = BSS.Process.Amber(ubiquitin_system, protocol)
 
     # Create a process to run with GROMACS.
-    process_gmx = BSS.Process.Gromacs(system, protocol, extra_options={"nsteps": 0})
+    process_gmx = BSS.Process.Gromacs(
+        ubiquitin_system, protocol, extra_options={"nsteps": 0}
+    )
 
     # Run the AMBER process and wait for it to finish.
     process_amb.start()
